rename the __init__.py file to argument_parser.py for examples folder
