+++ examples/history_vehicles_replacement_for_imitation_learning.py
@@ -2,7 +2,7 @@ import logging
 from dataclasses import replace
 
 from envision.client import Client as Envision
-from examples import default_argument_parser
+from examples.argument_parser import default_argument_parser
 from smarts.core.agent import Agent, AgentSpec
 from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.scenario import Mission, Scenario
