+++ examples/human_in_the_loop.py
@@ -4,7 +4,7 @@ import logging
 import gym
 from pynput.keyboard import Key, Listener
 
-from examples import default_argument_parser
+from examples.argument_parser import default_argument_parser
 from smarts.core.agent import Agent, AgentSpec
 from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.utils.episodes import episodes
