+++ examples/multi_instance.py
@@ -6,7 +6,7 @@ import numpy as np
 import ray
 import torch
 
-from examples import default_argument_parser
+from examples.argument_parser import default_argument_parser
 from smarts.core.agent import Agent, AgentSpec
 from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.utils.episodes import episodes
