+++ examples/single_agent.py
@@ -2,7 +2,7 @@ import logging
 
 import gym
 
-from examples import default_argument_parser
+from examples.argument_parser import default_argument_parser
 from smarts.core.agent import Agent, AgentSpec
 from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.sensors import Observation
